Composer trait for MainActivity
